+++ Changelog.md
@@ -3,6 +3,13 @@
 All notable changes to this project will be documented in this file.<br>
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.7.0] - 2025-04-21
+### Added
+- **[Camera2D]**: follow target and camera shake
+- **[Vector2]**: smoothDamp function
+### Fixed
+- **[Transform2D, ParticleRenderer2D]**: some sprites were not rendered when they should have been rendered
+
 ## [0.6.1] - 2025-04-20
 ### Added
 - **[Camera2D]**: 2D orthographic camera